Extract utils from reducer
